+++ examples/getCombinedLandsatSentinel2Wrapper.py
@@ -1,5 +1,5 @@
 """
-   Copyright 2022 Ian Housman
+   Copyright 2023 Ian Housman
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -130,9 +130,21 @@ cloudHeights = ee.List.sequence(500,10000,500)
 #This method works really well
 applyCloudProbability = True
 
+#Whether to use the pre-computed cloudScore+ to mask
+#clouds and cloud shadows for Sentinel 2
+#This method works really well and should be used instead of all other methods once it finishes
+# running for the S2 archive (~ early 2024????)
+applyCloudScorePlusSentinel2 = False
+
 #If cloudProbability is chosen, choose a threshold 
 #(generally somewhere around 40-60 works well)
 cloudProbThresh = 40
+
+# If applyCloudScorePlus = True, choose a threshold
+# Adjustable threshold for converting CS+ QA to a binary mask.
+# Higher thresholds will mask out more clouds (e.g. partial occlusions like thin clouds, haze &
+# cirrus shadows). Lower thresholds will mask out fewer clouds.
+cloudScorePlusThresh = 0.6
 
 #Fmask switches- only for Landsat
 #Generally we do use these
@@ -225,6 +237,12 @@ preComputedSentinel2TDOMIRStdDev = preComputedTDOMStats['sentinel2']['stdDev']
 # Whether to export composites
 exportComposites = False
 
+# Whether to overwite existing assets should they exist or currently be exporting
+overwrite = False
+
+# Whether to print a lot of information about what is going on
+verbose = False
+
 #Set up Names for the export
 outputName = 'Landsat_Sentinel2_Hybrid'
 
@@ -260,19 +278,18 @@ processedAndComposites = getImagesLib.getLandsatAndSentinel2HybridWrapper(studyA
   False,250,
   exportComposites,outputName,exportPathRoot,crs,transform,scale,
   preComputedLandsatCloudScoreOffset,preComputedLandsatTDOMIRMean,preComputedLandsatTDOMIRStdDev,preComputedSentinel2CloudScoreOffset,preComputedSentinel2TDOMIRMean,preComputedSentinel2TDOMIRStdDev,
-  cloudProbThresh,landsatCollectionVersion);
+  cloudProbThresh,landsatCollectionVersion,
+  overwrite, verbose, applyCloudScorePlusSentinel2, cloudScorePlusThresh)
 
 #Separate into scenes and composites for subsequent analysis
 processedScenes = processedAndComposites['processedScenes']
 processedComposites = processedAndComposites['processedComposites']
 
 # Indicate what type of image is being added to speed up map service creation
-getImagesLib.vizParamsFalse['layerType']= 'geeImage';
+getImagesLib.vizParamsFalse['layerType']= 'geeImage'
 
 # Map.addLayer(processedComposites.select(['NDVI','NBR']),{'addToLegend':'false'},'Time Series (NBR and NDVI)',False)
-for year in range(startYear + timebuffer      ,endYear + 1 - timebuffer ):
-     t = processedComposites.filter(ee.Filter.calendarRange(year,year,'year')).mosaic()
-     Map.addLayer(t.float(),getImagesLib.vizParamsFalse,str(year),False)
+Map.addTimeLapse(processedComposites,getImagesLib.vizParamsFalse,'Composite Timelapse',False)
 ####################################################################################################
 #Load the study region
 Map.addLayer(studyArea, {'strokeColor': '0000FF'}, "Study Area", True)